Add new data dictionary
Attached is a new data dictionary.  I have made the requested edits to the survey so there are now additional variables corresponding to each component of the score, and the value of these variables will be the value that is added to the score by the algorithm at the end of the MOCA.  I’ve also added variables to capture the date/time the MOCA is started, as well as the date and day of the week that the score is calculated.  In the attached spreadsheet, I’ve highlighted green in column A those rows which are new additions.  These changes are effective today 2/7/23 at 3:48 PM Eastern.