add the code that dealing with mutiple score issue
This code aims to retrieve records from a raw data source that have multiple scores and where the complete code(CO or 97 or 98) is not shown in the 'resCodeResult' variable. The code iterates over each record in the raw data and checks if any records have multiple scores by counting the number of entries in the 'scores' field. If a record has multiple scores, it then checks if the 'resCodeResult' variable contains the complete code. If the code is not found, the record is added to a list of results that satisfy both conditions. Finally, the code returns a CSV file of records that meet the criteria.